chore: migrate 404 page to App Router (#18449)
