+++ apps/web/test/lib/pagesAndRewritePaths.test.ts
@@ -23,7 +23,6 @@ describe("pagesAndRewritePaths", () => {
       "video",
       "workflows",
       "403",
-      "404",
       "500",
       "bookings",
       "event-types",
